+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Features**:
 
 - Add configuration to allow high cardinality tags in metrics. ([#4805](https://github.com/getsentry/relay/pull/4805))
+- Make client sdk mandatory for profile sample v2. ([#4853](https://github.com/getsentry/relay/pull/4853))
 
 **Internal**: